new storage config field & conversion methods
Added `convertFromStorage` and `convertToStorage`—doing what they say on the tin, but deferring to original value for zeroes and disabled sub-unit storage (which is the default). The latter is only called from the `process` method—feels a little like model accessors & mutators in stock Laravel. Within `getSubUnitFactor`, we have a default/fallback factor of 100.