asset cleanup
Renamed JS files to be less generic/more package-specific. Removed (commented out) CSS file, as it was empty. Also was causing a weird bug where Vite interpreted and built it as a JS file...not sure what that was about, but it caused issues in the manifest. Can see that under the "resources/css/cp.css" key in the manifest (it was being associated with a "...js" file.